+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Duck hunt".
 - Tests for "Condition overshadowing".
 - Tests for "Equalizing arrays".
+- Tests for "Level of nested parentheses".
 
 ## [1.15.0] - 2022-11-30
 ### Added